LayoutTests/fast/scrolling/rtl-scrollbars-overflow-elementFromPoint-expected.txt

This test makes sure that document.elementFromPoint() works correctly with RTL scrollbars.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.elementFromPoint(22, 7).id is "a"
PASS successfullyParsed is true

TEST COMPLETE
